+++ velox/exec/CMakeLists.txt
@@ -156,6 +156,7 @@ velox_add_library(
   IndexLookupJoin.h
   IndexLookupJoinBridge.h
   JoinBridge.h
+  KRangeFrameBound.h
   Limit.h
   LocalPartition.h
   LocalPlanner.h
@@ -181,6 +182,7 @@ velox_add_library(
   OutputBuffer.h
   OutputBufferManager.h
   ParallelProject.h
+  PeerGroupComputation.h
   PartitionFunction.h
   PartitionStreamingWindowBuild.h
   PartitionedOutput.h
@@ -229,7 +231,6 @@ velox_add_library(
   Window.h
   WindowBuild.h
   WindowFunction.h
-  WindowPartitionComputation.h
   WindowPartition.h
 )
